+++ openbas-api/src/main/java/io/openbas/rest/exercise/form/LessonsInput.java
@@ -7,6 +7,7 @@ import lombok.Setter;
 @Setter
 @Getter
 public class LessonsInput {
+    
     @JsonProperty("lessons_anonymized")
     private boolean lessonsAnonymized;
