+++ openbas-front/src/admin/components/simulations/simulation/lessons/ExerciseLessons.tsx
@@ -23,7 +23,7 @@ import type { TeamsHelper } from '../../../../../actions/teams/team-helper';
 import type { UserHelper } from '../../../../../actions/helper';
 import { fetchLessonsTemplates } from '../../../../../actions/Lessons';
 import { fetchPlayers } from '../../../../../actions/User';
-import { fetchExerciseInjects, fetchExerciseObjectives } from '../../../../../actions/Inject';
+import { fetchExerciseInjects } from '../../../../../actions/Inject';
 import { fetchExerciseTeams, updateExerciseLessons } from '../../../../../actions/Exercise';
 import Lessons from '../../../lessons/exercises/Lessons';
 import {
@@ -44,7 +44,7 @@ import {
 } from '../../../../../actions/exercises/exercise-action';
 import type { ExerciseStore } from '../../../../../actions/exercises/Exercise';
 import { usePermissions } from '../../../../../utils/Exercise';
-import { addExerciseObjective, deleteExerciseObjective, updateExerciseObjective } from '../../../../../actions/Objective';
+import { addExerciseObjective, deleteExerciseObjective, fetchExerciseObjectives, updateExerciseObjective } from '../../../../../actions/Objective';
 import { addExerciseEvaluation, fetchExerciseEvaluations, updateExerciseEvaluation } from '../../../../../actions/Evaluation';
 import { fetchTeams } from '../../../../../actions/teams/team-actions';
